Fixed a bug
Had a bug in my code that was causing duplication when a new note was entered...basically the optimistic state was sticking around with the real one (I guess that's how I'd put it?). After some digging I found a typo. After fixing the typo it appears to be working. Hopefully it stays that way...